# example/output/Bookstore/openapi.json
{"openapi": "3.0.0", "info": {"version": "0.0.1b", "title": "Bookstore", "description": ""}, "servers": [{"url": "http://localhost"}], "paths": {"/book": {"post": {"description": "Creates a new book.", "operationId": "createBook", "requestBody": {"description": "Book to add.", "required": true, "content": {"application/json": {"schema": {"$ref": "#components/schemas/Book"}}}}, "responses": {"200": {"$ref": "#components/responses/BookOK"}}}}, "/book/id": {"put": {"description": "Update Book object.", "operationId": "updateBook", "parameters": [{"name": "id", "in": "path", "description": "ID of book to update", "required": true, "schema": {"type": "string"}}], "requestBody": {"$ref": "#components/requestBodies/BookBody"}, "responses": {"200": {"$ref": "#components/responses/BookOK"}}}, "get": {"description": "Read Book object.", "operationId": "readBook", "parameters": [{"name": "id", "in": "path", "description": "ID of book to update", "required": true, "schema": {"type": "string"}}], "requestBody": {"$ref": "#components/requestBodies/BookBody"}, "responses": {"200": {"$ref": "#components/responses/BookOK"}}}, "delete": {"description": "Delete Book object.", "operationId": "deleteBook", "parameters": [{"name": "id", "in": "path", "description": "ID of book to update", "required": true, "schema": {"type": "string"}}], "requestBody": {"$ref": "#components/requestBodies/BookBody"}, "responses": {"200": {"$ref": "#components/responses/BookOK"}}}}, "/listbooks": {"get": {"description": "", "operationId": "listBooks", "parameters": [], "responses": {"200": {"description": "JSON representation of Book object", "content": {"application/json": {"schema": {"type": "array", "items": {"$ref": "#/components/schemas/Book"}}}}}}}}, "/bookexists/{isbn}": {"get": {"description": "", "operationId": "bookExists", "parameters": [{"name": "isbn", "in": "path", "required": true, "schema": {"type": "string"}}], "responses": {"200": {"description": "JSON representation of boolean object", "content": {"application/json": {"schema": {"type": "boolean"}}}}}}}, "/bookprice/{isbn}": {"get": {"description": "", "operationId": "bookPrice", "parameters": [{"name": "isbn", "in": "path", "required": true, "schema": {"type": "string"}}], "responses": {"200": {"description": "JSON representation of number object", "content": {"application/json": {"schema": {"type": "number"}}}}}}}}, "components": {"schemas": {"Book": {"type": "object", "properties": {"isbn": {"type": "string"}, "title": {"type": "string"}, "author": {"type": "string"}, "category": {"type": "string"}, "price": {"type": "number", "format": "double"}}, "required": ["title", "author", "price"]}}, "requestBodies": {"BookBody": {"description": "JSON representation of Book object", "required": true, "content": {"application/json": {"schema": {"$ref": "#components/schemas/Book"}}}}}, "responses": {"BookOK": {"description": "JSON representation of Book object", "content": {"application/json": {"schema": {"$ref": "#components/schemas/Book"}}}}}}}